+++ src/main/java/net/silentchaos512/gear/compat/jei/ToolActionRecipeCategory.java
@@ -3,11 +3,13 @@ package net.silentchaos512.gear.compat.jei;
 import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
+import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
 import mezz.jei.api.helpers.IGuiHelper;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.recipe.category.IRecipeCategory;
+import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
@@ -35,6 +37,16 @@ public class ToolActionRecipeCategory implements IRecipeCategory<ToolActionRecip
     }
 
     @Override
+    public int getWidth() {
+        return GUI_WIDTH;
+    }
+
+    @Override
+    public int getHeight() {
+        return GUI_HEIGHT;
+    }
+
+    @Override
     public RecipeType<ToolActionRecipe> getRecipeType() {
         return SGearJeiPlugin.TOOL_ACTION_TYPE;
     }
@@ -42,11 +54,6 @@ public class ToolActionRecipeCategory implements IRecipeCategory<ToolActionRecip
     @Override
     public Component getTitle() {
         return localizedName;
-    }
-
-    @Override
-    public IDrawable getBackground() {
-        return background;
     }
 
     @Override
@@ -64,5 +71,10 @@ public class ToolActionRecipeCategory implements IRecipeCategory<ToolActionRecip
                 .addIngredients(Ingredient.of(SgBlocks.STONE_ANVIL.get()));
         builder.addSlot(RecipeIngredientRole.OUTPUT, 76, 12)
                 .addIngredients(Ingredient.of(recipe.getResult()));
+    }
+
+    @Override
+    public void draw(ToolActionRecipe recipe, IRecipeSlotsView recipeSlotsView, GuiGraphics guiGraphics, double mouseX, double mouseY) {
+        background.draw(guiGraphics);
     }
 }