+++ src/main/java/net/silentchaos512/gear/crafting/recipe/salvage/SalvagingRecipe.java
@@ -7,6 +7,7 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.world.Container;
+import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.*;
 import net.minecraft.world.level.Level;
@@ -35,6 +36,10 @@ public class SalvagingRecipe implements Recipe<SingleRecipeInput> {
 
     public List<ItemStack> getPossibleResults(Container inv) {
         return new ArrayList<>(results);
+    }
+
+    public List<ItemStack> getPossibleResultsForDisplay() {
+        return getPossibleResults(new SimpleContainer(1));
     }
 
     @Override